lamp_stack/requirements.txt
pyserial==3.5
mysql-connector-python==8.2.0
matplotlib==3.8.4
numpy==1.24.3
